+++ sdk/keyvault/azure-keyvault-certificates/azure/keyvault/certificates/aio/client.py
@@ -783,7 +783,7 @@ class CertificateClient(AsyncKeyVaultClientBase):
         :keyword str organization_id: Id of the organization
         :keyword admin_details: Details of the organization administrators of the
          certificate issuer.
-        :paramtype admin_details: list[~azure.keyvault.certificates.models.AdministratorDetails]
+        :paramtype admin_details: list[~azure.keyvault.certificates.models.AdministratorContact]
         :returns: The created CertificateIssuer
         :rtype: ~azure.keyvault.certificates.models.CertificateIssuer
         :raises: :class:`~azure.core.exceptions.HttpResponseError`
@@ -854,7 +854,7 @@ class CertificateClient(AsyncKeyVaultClientBase):
         :keyword str password: The password/secret/account key.
         :keyword str organization_id: Id of the organization
         :keyword admin_details: Details of the organization administrators of the certificate issuer
-        :paramtype admin_details: list[~azure.keyvault.certificates.models.AdministratorDetails]
+        :paramtype admin_details: list[~azure.keyvault.certificates.models.AdministratorContact]
         :return: The updated issuer
         :rtype: ~azure.keyvault.certificates.models.CertificateIssuer
         :raises: :class:`~azure.core.exceptions.HttpResponseError`